+++ src/routes/products.ts
@@ -8,7 +8,6 @@ import { ParamsDictionary } from 'express-serve-static-core';
 import { body, validationResult } from 'express-validator';
 import { NO_CONTENT } from 'http-status';
 import * as moment from 'moment-timezone';
-import * as _ from 'underscore';
 
 import * as Message from '../message';
 
@@ -81,7 +80,6 @@ productsRouter.all<any>(
                 accounting: {}
             },
             itemOffered: { name: {} },
-            seatReservationUnit: (_.isEmpty(req.body.seatReservationUnit)) ? 1 : req.body.seatReservationUnit,
             ...req.body
         };
         if (forms.additionalProperty.length < NUM_ADDITIONAL_PROPERTY) {