+++ CHANGELOG.md
@@ -18,6 +18,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Security
 
+## v2.20.1 - 2020-12-24
+
+### Changed
+
+- update @chevre/api-nodejs-client
+- update @cinerino/sdk
+- uninstall underscore
+- アクション検索条件拡張
+
 ## v2.20.0 - 2020-12-18
 
 ### Added